Widen relay seq to 12 bits and move it into the payload
Random 8-bit nonces collided too often in the dedup ring. Widen the sequence
number to 12 bits using the 4 spare payload bits, and move it out of the
manufacturer-data header into the payload (bits 116-127) so the mesh and macOS
paths carry and read it identically.

- Header drops its seq byte (6->5 bytes); seq now lives in the payload via
  message_get_seq/message_set_seq (12-bit).
- Mesh originate goes back to an incrementing counter (g_seq, wraps at 0xFFF)
  so back-to-back triggers never collide. Dedup is unchanged — still a pure
  existence check, making no assumption the counter increases, so a reset that
  restarts it at 0 is harmless beyond a brief stale-entry collision chance.
- The two RX paths share a new dedup_and_act() helper; the macOS gateway packet
  no longer writes a header seq byte.
- broadcast_log seq widened uint8_t -> uint16_t through core/wrapper/headers.
- bleatctl.py mirrors the 12-bit payload seq (random nonce — the Mac is
  stateless per run and dedup only checks existence).
- tests: 12-bit seq round-trip, no bleed into slots/command, wide-seq dedup.

Breaking wire change — the whole fleet must run this firmware together.

+++ app/src/broadcast_log.c
@@ -17,7 +17,7 @@ void broadcast_log_init(void)
     k_spin_unlock(&g_lock, key);
 }
 
-bool broadcast_log_check_and_record(uint8_t originator, uint8_t seq)
+bool broadcast_log_check_and_record(uint8_t originator, uint16_t seq)
 {
     k_spinlock_key_t key = k_spin_lock(&g_lock);
     bool duplicate = broadcast_log_core_check_and_record(&g_log,
@@ -26,7 +26,7 @@ bool broadcast_log_check_and_record(uint8_t originator, uint8_t seq)
     return duplicate;
 }
 
-void broadcast_log_record(uint8_t originator, uint8_t seq)
+void broadcast_log_record(uint8_t originator, uint16_t seq)
 {
     k_spinlock_key_t key = k_spin_lock(&g_lock);
     broadcast_log_core_record(&g_log, originator, seq);
